fixed fmriprep directory unzipping procedure

+++ run.sh
@@ -31,21 +31,24 @@ else
 fi
 
 # Show the contents of the BIDS directory
-ls -R "${BIDS_DIR}"
+BIDS_DIR=${INPUT_DIR}/bids_dataset
+ls -R ${BIDS_DIR}
 
 # Position fmriprepdir contents
-unzip ${INPUT_DIR}/fmriprepdir -d ${INPUT_DIR}
+unzip ${INPUT_DIR}/fmriprepdir/*.zip -d ${INPUT_DIR}
 cd ${INPUT_DIR} || exit
+if [ ! -d ${INPUT_DIR}/fmriprepdir ]; then
 rm -rf ${INPUT_DIR}/fmriprepdir
-find . -maxdepth 2 -type d | grep -E -v bids | grep -E -v fmriprepdir | grep -E fmriprep
+fi
+FMRIPREP_DIR=$(find $(pwd) -maxdepth 2 -type d | grep -E -v bids | grep -E -v fmriprepdir | grep -E fmriprep)
 cd ${FLYWHEEL_BASE} || exit
 
 # Copy event files
-cp ${FLYWHEEL_BASE}/events/* ${FLYWHEEL_BASE}/input/bids_dataset/
+cp ${FLYWHEEL_BASE}/events/* ${INPUT_DIR}/bids_dataset/
 
 # Run script
-/usr/local/miniconda/bin/python3 report.py ${INPUT_DIR}/bids_dataset ${INPUT_DIR}/fmriprepdir ${RESULTS_DIR}
+/usr/local/miniconda/bin/python3 report.py "${BIDS_DIR}" "${FMRIPREP_DIR}" "${RESULTS_DIR}"
 
 # Position results directory as zip file in /flywheel/v0/output
 zip -r report_results.zip report_results
-cp report_results.zip ${OUTPUT_DIR}/
+mv report_results.zip ${OUTPUT_DIR}/
